pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "nedbank-mlops"
version = "0.1.0"
description = "MLOps skeleton for Nedbank Data & Analytics Masters Challenge"
requires-python = ">=3.10"

dependencies = [
    # Core data
    "pandas>=2.1",
    "polars>=0.20",
    "numpy>=1.26",
    "pyarrow>=14.0",

    # ML
    "scikit-learn>=1.4",
    "lightgbm>=4.3",
    "xgboost>=2.0",
    "imbalanced-learn>=0.12",
    "optuna>=3.5",

    # MLOps / tracking
    "mlflow>=2.10",
    "dvc>=3.40",
    "dvc-s3>=3.0",

    # Config management
    "hydra-core>=1.3",
    "omegaconf>=2.3",

    # Data validation
    "great-expectations>=0.18",
    "pandera>=0.18",

    # Observability / DE
    "pydantic>=2.5",
    "structlog>=24.1",
    "prometheus-client>=0.20",

    # Utilities
    "typer>=0.12",
    "rich>=13.7",
    "python-dotenv>=1.0",
    "joblib>=1.3",
    "tqdm>=4.66",
    "scipy>=1.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.3",
    "mypy>=1.9",
    "pre-commit>=3.7",
    "ipykernel>=6.29",
    "jupyter>=1.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "plotly>=5.20",
    "shap>=0.45",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
